refactor: storybook 관련 유틸 .storybook 폴더 내부로 이동
